auth_api/python/requirements.txt
Flask
PyJWT
sqlalchemy
mysqlclient
